requirements.txt
numpy==1.21.6
pandas==1.3.5
protobuf==4.21.6
pybigquery==0.10.2
SQLAlchemy==1.4.41
seaborn
matplotlib
google-cloud-bigquery